+++ ui/user/src/routes/admin/users/+page.ts
@@ -4,5 +4,5 @@ import { redirect } from '@sveltejs/kit';
 export const load: PageLoad = ({ url }) => {
 	const searchParams = new URLSearchParams(url.searchParams);
 	searchParams.delete('view');
-	throw redirect(301, `/admin/identity-access?view=users&${searchParams}`);
+	throw redirect(301, `/identity-access?view=users&${searchParams}`);
 };
