+++ ui/user/src/routes/identity-access/AddGroupAssignmentDialog.svelte
