+++ ui/user/src/routes/identity-access/AssignGroupRoleDialog.svelte
