+++ ui/user/src/routes/identity-access/ConfirmAuditorRoleDialog.svelte
