+++ ui/user/src/routes/identity-access/ConfirmOwnerRoleDialog.svelte
