+++ ui/user/src/routes/identity-access/ConfirmUserImpersonationRoleDialog.svelte
