+++ ui/user/src/routes/identity-access/agents/[id]/page.svelte.spec.ts
@@ -1,6 +1,6 @@
-import { ApiKeysService } from '$lib/services';
+import { ApiKeysService, Group } from '$lib/services';
 import type { APIKey } from '$lib/services/api-keys/types';
-import { createMockProfile } from '../../../tests/helpers/pageData';
+import { createMockProfile } from '../../../../tests/helpers/pageData';
 import { load } from './+page';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
@@ -15,13 +15,12 @@ const apiKey: APIKey = {
 	createdAt: '2026-01-01T00:00:00.000Z'
 };
 
-function loadAgentAuthScope(pathname: string) {
-	const profile = createMockProfile();
+function loadAgentAuthScope(groups: string[]) {
+	const profile = createMockProfile(groups);
 	return load({
 		fetch: vi.fn(),
 		params: { id: apiKey.id.toString() },
-		parent: vi.fn(async () => ({ profile })),
-		url: new URL(pathname, 'http://localhost')
+		parent: vi.fn(async () => ({ profile }))
 	} as unknown as Parameters<typeof load>[0]);
 }
 
@@ -29,23 +28,23 @@ afterEach(() => {
 	vi.restoreAllMocks();
 });
 
-describe('Agent Auth Scope detail route selection', () => {
-	it('gets current-user scope for non-admin URL', async () => {
+describe('Agent Auth Scope detail load', () => {
+	it('gets the current-user scope for users without admin access', async () => {
 		const getApiKey = vi.spyOn(ApiKeysService, 'getApiKey').mockResolvedValue(apiKey);
 		const getAnyApiKey = vi.spyOn(ApiKeysService, 'getAnyApiKey');
 
-		const result = await loadAgentAuthScope(`/agent-auth-scopes/${apiKey.id}`);
+		const result = await loadAgentAuthScope([Group.USER]);
 
 		expect(result).toEqual({ apiKey, isAdmin: false });
 		expect(getApiKey).toHaveBeenCalledWith(apiKey.id.toString(), expect.any(Object));
 		expect(getAnyApiKey).not.toHaveBeenCalled();
 	});
 
-	it('gets any scope for admin URL', async () => {
+	it('gets any scope for users with admin access', async () => {
 		const getApiKey = vi.spyOn(ApiKeysService, 'getApiKey');
 		const getAnyApiKey = vi.spyOn(ApiKeysService, 'getAnyApiKey').mockResolvedValue(apiKey);
 
-		const result = await loadAgentAuthScope(`/admin/agent-auth-scopes/${apiKey.id}`);
+		const result = await loadAgentAuthScope([Group.ADMIN]);
 
 		expect(result).toEqual({ apiKey, isAdmin: true });
 		expect(getApiKey).not.toHaveBeenCalled();
